Move LinearMapInfo out to a separate file.
