+++ lib/SILOptimizer/Utils/Differentiation/ADContext.cpp
@@ -18,7 +18,6 @@
 #define DEBUG_TYPE "differentiation"
 
 #include "swift/AST/DiagnosticsSIL.h"
-#include "swift/AST/Expr.h"
 #include "swift/SILOptimizer/PassManager/Transforms.h"
 #include "swift/SILOptimizer/Utils/Differentiation/ADContext.h"
