+++ lib/SILOptimizer/Utils/Differentiation/CMakeLists.txt
@@ -3,4 +3,5 @@ silopt_register_sources(
   Common.cpp
   DerivativeLookup.cpp
   DifferentiationInvoker.cpp
+  LinearMapInfo.cpp
 )
